Updated QFJPage to use Scripture dict and dKey
Started cleaning up the code by creating an array of dictionary keys
(dKey) to index the updated Scripture dictionary (instead of string
dictionary).

I upgraded the Xamarin.Forms package in Evangelizer, and in .iOS and
.Android.

And then (last night) I was unable to build and execute in either the
iOS or the Android simulator as it was giving me the "No Embedded
Resources Found" error message.

I tried again this morning and the build deployed to iOS Simulator and
to Genymotion (Android sim) and they both worked fine.